boot0img: allow patching of boot0 to put U-Boot below 1MB
boot0 loads the binaries from 19096 KB of the SD card.
Add the "-B" option to patch boot0 to load the firmware from 40K instead,
which is just after boot0 itself.
This scans the boot0 binary for the two "movw" Thumb2 instructions with the
original sector value and replaces it with the new offset.