binaries: add FEL capable SPL binaries
FEL booting allows to boot a board using an USB OTG connection from a
host PC, without the need to use any media (SD card) or on-board storage
(SPI flash). This is controlled by the BootROM, which runs entirely in
AArch32 mode. To allow FEL booting, we need to load and execute some code
to at least initialise the DRAM. U-Boot's SPL can do this, but it needs
to return to the (32 bit!) BootROM to allow loading of further firmware
components or other images.
Since the mainline U-Boot is an AArch64 port, this is not easily doable
with the upstream version. However it is possible to build an AArch32
version of the SPL, with the help of some (non-upstream) patches [1].

Add three binaries to allow FEL booting of A64 (DDR3/LPDDR3 DRAM) and
H5 boards.
Also document the steps required to rebuild those binaries.

[1] https://github.com/apritzel/u-boot/commits/sunxi64-fel32