binaries: Update mainline TF-A binary files to v2.4
Update the almost two year old versions of the Trusted Firmware binaries
with an up-to-date version.
This is from the new v2.4 release, featuring advanced PMIC and DT
capabilities, and including support for the "crust" AR100 firmware.

Move the old files into an "obsolete" directory, and update the README
on the way to reflect the ATF => TF-A rename.